+++ client/src/components/molecules/index.tsx
@@ -1,5 +1,7 @@
 export { default as IconButton } from 'components/molecules/IconButton';
 export { default as TextButton } from 'components/molecules/TextButton';
+export { default as PrimaryButton } from 'components/molecules/PrimaryButton';
 export { default as MindmapBackground } from 'components/molecules/MindmapBackground';
-export { default as MindmapTree } from 'components/molecules/MindmapTree';
 export { default as Header } from 'components/molecules/Header';
+export { default as Path } from 'components/molecules/Path';
+export { default as Tree } from 'components/molecules/Tree';
