+++ client/src/components/organisms/MindmapBtnWrapper/index.tsx
@@ -1,9 +1,9 @@
 import styled from '@emotion/styled';
 import { useHistory } from 'react-router-dom';
 import { clock, plusCircle } from 'img';
-import PrimaryButton from 'components/molecules/PrimaryButton';
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { mindmapState, selectedNodeState } from 'recoil/mindmap';
+import { IMindNodes, mindNodesState, selectedNodeState } from 'recoil/mindmap';
+import { PrimaryButton } from 'components/molecules';
 import { getId, idxToLevel, levelToIdx } from 'utils/helpers';
 
 const Wrapper = styled.div`
@@ -12,37 +12,41 @@ const Wrapper = styled.div`
   position: fixed;
   bottom: 50px;
   left: 50px;
-  > *:not(:last-child) {
-    margin-right: 10px;
-  }
+  gap: 1rem;
 `;
 
-const useHistoryBtnClick = () => {
+interface ITempNodeParams {
+  mindNodes: IMindNodes;
+  selectedNodeId: string | null;
+}
+
+const createTempNode = ({ mindNodes, selectedNodeId }: ITempNodeParams) => {
+  const nodeId = selectedNodeId === null ? 0 : getId(selectedNodeId);
+  const parentNode = mindNodes.get(nodeId);
+  const level = idxToLevel(levelToIdx(parentNode!.level) + 1);
+  const tempNode = { nodeId: -1, level: level, content: '', children: [] };
+
+  parentNode!.children.push(-1);
+  mindNodes.set(-1, tempNode);
+
+  const tempNodeId = `${level}#-1`;
+  return tempNodeId;
+};
+
+const MindmapWrapper: React.FC = () => {
+  const [selectedNodeId, setSelectedNodeId] = useRecoilState(selectedNodeState);
+  const mindNodes = useRecoilValue(mindNodesState);
   const hitory = useHistory();
 
   const handleHistoryBtnClick = () => {
     hitory.push('/history');
   };
 
-  return handleHistoryBtnClick;
-};
-
-const MindmapWrapper = () => {
-  const [selectedNodeId, setSelectedNodeId] = useRecoilState(selectedNodeState);
-
-  const handleHistoryBtnClick = useHistoryBtnClick();
-  const mindNodes = useRecoilValue(mindmapState).mindNodes;
-
   const handlePlusNodeBtnClick = () => {
     if (selectedNodeId && getId(selectedNodeId) === -1) return;
-    const nodeId = selectedNodeId === null ? 0 : getId(selectedNodeId);
-    const parentNode = mindNodes.get(nodeId);
-    const parentNodeIdx = levelToIdx(parentNode!.level);
-    const level = idxToLevel(parentNodeIdx + 1);
-    const tempNode = { nodeId: -1, level: level, content: '', children: [] };
-    parentNode?.children.push(-1);
-    mindNodes.set(-1, tempNode);
-    setSelectedNodeId(`${level}#-1`);
+
+    const tempNodeId = createTempNode({ mindNodes, selectedNodeId });
+    setSelectedNodeId(tempNodeId);
   };
 
   return (
