+++ client/src/components/organisms/index.tsx
@@ -5,3 +5,4 @@ export { default as ProjectCard } from 'components/organisms/ProjectCard';
 export { default as TextInputModal } from 'components/organisms/TextInputModal';
 export { default as MindmapBtnWrapper } from 'components/organisms/MindmapBtnWrapper';
 export { default as Mindmap } from 'components/organisms/Mindmap';
+export { default as MindmapTree } from 'components/organisms/MindmapTree';
