+++ client/src/components/templates/MindmapTemplate/index.tsx
@@ -3,15 +3,13 @@ import { mindmapState } from 'recoil/mindmap';
 import { MindmapBackground } from 'components/molecules';
 import { Mindmap, MindmapBtnWrapper } from 'components/organisms';
 
-const MindmapTemplate = () => {
+const MindmapTemplate: React.FC = () => {
   const mindmapData = useRecoilValue(mindmapState);
   return (
-    <>
-      <MindmapBackground className='mindmap-area'>
-        <Mindmap mindmapData={mindmapData} />
-        <MindmapBtnWrapper />
-      </MindmapBackground>
-    </>
+    <MindmapBackground className='mindmap-area'>
+      <Mindmap mindmapData={mindmapData} />
+      <MindmapBtnWrapper />
+    </MindmapBackground>
   );
 };
