+++ client/src/recoil/mindmap/index.ts
@@ -1,4 +1,4 @@
-import { atom } from 'recoil';
+import { atom, selector } from 'recoil';
 import { Levels } from 'utils/helpers';
 
 export interface IMindmapData {
@@ -59,6 +59,13 @@ export const mindmapState = atom<IMindmapData>({
   key: 'mindmapAtom',
   // default: { rootId: initRootId, mindNodes: new Map([[initRootId, initRootNode]]) },
   default: getDummyMindmapData(),
+});
+
+export const mindNodesState = selector({
+  key: 'mindNodesState',
+  get: ({ get }) => {
+    return get(mindmapState).mindNodes;
+  },
 });
 
 export const selectedNodeState = atom<string | null>({ key: 'selectedNodeAtom', default: null });
